extract integritee bridge setup to its own module
